move publish to docker quickpick code to its own file
